Loading times dramatically improved by using an integer handle, also less stragglers in memory.  1GB mem usage to about 60MB.  Terrain generators also optimized to hell and back.

+++ NBTForgeCS/UIControls/MapControl.cs
@@ -332,7 +332,7 @@ namespace MineEdit
                     new MenuItem("Replace..."),//,new EventHandler(delegate(object s,EventArgs ea){})),
                     new MenuItem("Paint..."),//,new EventHandler(delegate(object s,EventArgs ea){})),
                     new MenuItem("Generate...",new EventHandler(delegate(object s,EventArgs ea){
-                        int min, max;
+                        double min, max;
                         Map.Generate(Map, mcc.AssignedChunk.X,mcc.AssignedChunk.Y, out min, out max);
                         Map.LoadChunk(mcc.AssignedChunk.X, mcc.AssignedChunk.Y);
                         mcc.Render();
